basic Server controlled position, some multithreading

+++ Project2/Project2/Player.h
@@ -31,7 +31,7 @@ public:
 
 
 
-	void Update(double dt){
+	void Update(long dt){
 		if (Entity::isMoving()){
